Introduce distinction between preview/productive site generation
Posts which contain the YAML attribute `published: false` are usually
not generated by Jekyll. With this patch they can be previewed just
like published posts on localhost using `rake watch`or `rake preview`.

NOTICE: Before pushing to the productive environment, use `rake generate`
to update the public directory and remove posts which are flagged not to be
published.

Former-commit-id: f1ebf358693582d310aac2599870c197a5188e8f